requirements.txt
Pillow
streamlit == 0.84.0
opencv-python
transformer
torch